# newsradar/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NewsRadar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/site/editions": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "List Editions",
        "operationId": "list_editions_site_editions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EditionSummaryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/site/editions/{edition_id}": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Edition",
        "operationId": "get_edition_site_editions__edition_id__get",
        "parameters": [
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Edition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/site/story/{story_type}/{story_id}": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Story",
        "operationId": "get_story_site_story__story_type___story_id__get",
        "parameters": [
          {
            "name": "story_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Story Type"
            }
          },
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Story Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/site/story/event/{event_id}/full-coverage": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "Get Full Coverage",
        "description": "Google-News \"Full Coverage\" for an event: coverage grouped into angles\n(embedding sub-clusters) plus by-country and entity-targeted stance facets.\n404 when the event is missing or has fewer than two distinct outlets.",
        "operationId": "get_full_coverage_site_story_event__event_id__full_coverage_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullCoverageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/share-links": {
      "get": {
        "tags": [
          "site"
        ],
        "summary": "List Share Links",
        "operationId": "list_share_links_share_links_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShareLinkOut"
                  },
                  "type": "array",
                  "title": "Response List Share Links Share Links Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "site"
        ],
        "summary": "Create Share",
        "operationId": "create_share_share_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareLinkCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/share-links/{link_id}/revoke": {
      "post": {
        "tags": [
          "site"
        ],
        "summary": "Revoke Share",
        "operationId": "revoke_share_share_links__link_id__revoke_post",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/share-links/{link_id}": {
      "delete": {
        "tags": [
          "site"
        ],
        "summary": "Delete Share",
        "operationId": "delete_share_share_links__link_id__delete",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Link Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/p/{token}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Public View",
        "description": "Serve the shared edition/interest feed (or the latest digest for a digest link).",
        "operationId": "public_view_p__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EditionOut"
                    },
                    {
                      "type": "object",
                      "additionalProperties": true
                    }
                  ],
                  "title": "Response Public View P  Token  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountryCoverageOut": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "country",
          "count"
        ],
        "title": "CountryCoverageOut",
        "description": "Outlet count for one source country (where the outlet is based)."
      },
      "CoverageAngleOut": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "outlets": {
            "items": {
              "$ref": "#/components/schemas/CoverageItemOut"
            },
            "type": "array",
            "title": "Outlets"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "earliest": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest"
          }
        },
        "type": "object",
        "required": [
          "label",
          "outlets",
          "size",
          "earliest"
        ],
        "title": "CoverageAngleOut",
        "description": "One angle (framing sub-cluster) of an event's coverage \u2014 the Google-News\n\"Full Coverage\" grouping. Outlets in an angle framed the story similarly\n(grouped by document-embedding similarity). ``label`` is the angle's\nrepresentative headline (translated when available); attribution-only, no body."
      },
      "CoverageItemOut": {
        "properties": {
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "source_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Country"
          }
        },
        "type": "object",
        "required": [
          "source_name",
          "url",
          "published_at",
          "source_country"
        ],
        "title": "CoverageItemOut",
        "description": "One outlet's coverage of an event-backed story (attribution, sorted by time)."
      },
...
      "EditionItemOut": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "story": {
            "$ref": "#/components/schemas/StoryOut"
          }
        },
        "type": "object",
        "required": [
          "position",
          "section",
          "story"
        ],
        "title": "EditionItemOut"
      },
      "EditionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "lookback_hours": {
            "type": "integer",
            "title": "Lookback Hours"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          },
          "sections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sections"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EditionItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "generated_at",
          "lookback_hours",
          "item_count"
        ],
        "title": "EditionOut"
      },
      "EditionSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "lookback_hours": {
            "type": "integer",
            "title": "Lookback Hours"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "generated_at",
          "lookback_hours",
          "item_count"
        ],
        "title": "EditionSummaryOut"
      },
...
      "FullCoverageOut": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "total_outlets": {
            "type": "integer",
            "title": "Total Outlets"
          },
          "angles": {
            "items": {
              "$ref": "#/components/schemas/CoverageAngleOut"
            },
            "type": "array",
            "title": "Angles"
          },
          "by_country": {
            "items": {
              "$ref": "#/components/schemas/CountryCoverageOut"
            },
            "type": "array",
            "title": "By Country"
          },
          "stance": {
            "$ref": "#/components/schemas/StanceSummaryOut"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "total_outlets",
          "angles",
          "by_country",
          "stance"
        ],
        "title": "FullCoverageOut",
        "description": "Full-coverage breakdown for an event-backed story (P8): coverage grouped\ninto angles plus by-country and entity-targeted stance facets. Every outlet\ncarries attribution (name + original url); no body text is included."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_EditionSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EditionSummaryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[EditionSummaryOut]"
      },
...
      "ShareLinkCreateIn": {
        "properties": {
          "scope": {
            "type": "string",
            "pattern": "^(site|edition|interest|digest)$",
            "title": "Scope"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "scope"
        ],
        "title": "ShareLinkCreateIn"
      },
      "ShareLinkOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "token",
          "scope",
          "target_id",
          "label",
          "expires_at",
          "revoked_at",
          "view_count",
          "created_at"
        ],
        "title": "ShareLinkOut"
      },
...
      "StoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "story_type": {
            "type": "string",
            "title": "Story Type"
          },
          "headline_en": {
            "type": "string",
            "title": "Headline En"
          },
          "headline_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline Original"
          },
          "source_lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Lang"
          },
          "translation_status": {
            "type": "string",
            "title": "Translation Status"
          },
          "extract_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extract En"
          },
          "body_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body En"
          },
          "blurb": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blurb"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Domain"
          },
          "source_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Country"
          },
          "favicon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favicon Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "image_alt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Alt"
          },
          "byline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Byline"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "frameable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frameable"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "personal_score": {
            "type": "number",
            "title": "Personal Score"
          },
          "coverage": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CoverageItemOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "story_type",
          "headline_en",
          "headline_original",
          "source_lang",
          "translation_status",
          "extract_en",
          "blurb",
          "source_name",
          "source_domain",
          "source_country",
          "favicon_url",
          "image_url",
          "image_alt",
          "byline",
          "published_at",
          "url",
          "frameable",
          "reason",
          "personal_score"
        ],
        "title": "StoryOut",
        "description": "The single reader story payload. Built ONLY by ``site.serializers.to_story_out``.\n\n``body_en`` is present only for ``full_ok`` sources; ``extract_en`` is capped at\nthe source's content-rights length. ``source_name`` and ``url`` (the original\narticle) are always populated (the attribution gate)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}